# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateIssuedInvoiceRequest:
    properties:
      amount:
        type: number
      customerId:
        type: string
      notes:
        type: string
      status:
        type: string
    type: object
...
  handler.IssuedInvoiceResponse:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      customerId:
        type: string
      id:
        type: string
      notes:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.PatchIssuedInvoiceRequest:
    properties:
      amount:
        type: number
      notes:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API de gestión de taller: autenticación JWT, coches, citas, empleados,
    proveedores, facturas recibidas/emitidas y documentos de facturación.'
  title: GonsGarage API
  version: "1.0"
paths:
...
  /api/v1/invoices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{items,total}'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar todas las facturas emitidas (taller)
      tags:
      - invoices
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateIssuedInvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.IssuedInvoiceResponse'
      security:
      - BearerAuth: []
      summary: Crear factura emitida a cliente
      tags:
      - invoices
  /api/v1/invoices/{id}:
    delete:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Eliminar factura emitida
      tags:
      - invoices
    get:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IssuedInvoiceResponse'
      security:
      - BearerAuth: []
      summary: Obtener factura emitida al cliente
      tags:
      - invoices
    patch:
      consumes:
      - application/json
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        type: string
      - description: Campos
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PatchIssuedInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IssuedInvoiceResponse'
      security:
      - BearerAuth: []
      summary: 'Actualizar factura (notas cliente; staff: estado/importe)'
      tags:
      - invoices
...
  /api/v1/invoices/me:
    get:
      parameters:
      - description: Límite (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{items,total}'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar mis facturas (cliente)
      tags:
      - invoices
...
securityDefinitions:
  BearerAuth:
    description: 'JWT: cabecera Authorization con valor Bearer seguido del token (rutas
      bajo /api/v1 salvo /health).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"